Rewrite unnest bool ops via expr helper
