+++ src/transform/mod.rs
@@ -2,6 +2,7 @@ pub(crate) mod context;
 pub(crate) mod expr;
 pub(crate) mod rewrite_assert;
 pub(crate) mod rewrite_class_def;
+pub(crate) mod rewrite_complex_expr;
 pub(crate) mod rewrite_decorator;
 pub(crate) mod rewrite_expr_to_stmt;
 pub(crate) mod rewrite_for_loop;
